HT_20/requirements.txt
Django~=4.0.2
Pillow~=9.0.1
djangorestframework~=3.13.1